Copied in the windows files to revert the spacing to make later merges easier Added in some more STUBS to make merging easier Removed all Warnings

+++ Src/BeebWin.h
@@ -164,7 +164,7 @@ public:
 
 	LRESULT WndProc(UINT nMessage, WPARAM wParam, LPARAM lParam);
 #endif
-	
+
 	void UpdateModelMenu();
 	void SetSoundMenu(void);
 	void SetImageName(const char *DiscName, int Drive, DiscType Type);
@@ -189,7 +189,7 @@ public:
 #ifndef __APPLE__
 		int d = (y*800)+sx+ScreenAdjust+(TeletextEnabled?36:0);
 #else
-		int d = (int)((y*800)+sx+ScreenAdjust+(TeletextEnabled?36:0));
+		long d = (y*800)+sx+ScreenAdjust+(TeletextEnabled?36:0);
 #endif
 		if ((d+width)>(500*800)) return;
 		if (d<0) return;
@@ -201,7 +201,7 @@ public:
 #ifndef __APPLE__
 		int d = (y*800)+sx+ScreenAdjust+(TeletextEnabled?36:0);
 #else
-		int d = (int)((y*800)+sx+ScreenAdjust+(TeletextEnabled?36:0));
+		long d = (y*800)+sx+ScreenAdjust+(TeletextEnabled?36:0);
 #endif
 		char *vaddr;
 		if ((d+width)>(500*800)) return;
@@ -220,7 +220,7 @@ public:
 #ifndef __APPLE__
 		int d = (y*800)+ScreenAdjust;
 #else
-		int d = (int)((y*800)+ScreenAdjust);
+		long d = (y*800)+ScreenAdjust;
 #endif
 		if (d > (MAX_VIDEO_SCAN_LINES*800))
 			return((EightUChars *)(m_screen+(MAX_VIDEO_SCAN_LINES*800)));
@@ -231,7 +231,7 @@ public:
 #ifndef __APPLE__
 		int d = (y*800)+ScreenAdjust;
 #else
-		int d = (int)((y*800)+ScreenAdjust);
+		long d = (y*800)+ScreenAdjust;
 #endif
 		if (d > (MAX_VIDEO_SCAN_LINES*800))
 			return((SixteenUChars *)(m_screen+(MAX_VIDEO_SCAN_LINES*800)));
@@ -449,7 +449,7 @@ public:
 	bool RegGetStringValue(HKEY hKeyRoot, LPCSTR lpSubKey, LPCSTR lpValue, LPSTR pData, DWORD dwSize);
 	bool RegSetStringValue(HKEY hKeyRoot, LPCSTR lpSubKey, LPCSTR lpValue, LPCSTR pData);
 #endif
-	
+
 	// Preferences
 	void LoadPreferences();
 	void SavePreferences(bool saveAll);
@@ -555,7 +555,7 @@ public:
 	LPDIRECT3DTEXTURE9 m_pTexture;
 	D3DXMATRIX m_TextureMatrix;
 #endif
-	
+
 	// Audio
 	UINT m_MenuIDSampleRate;
 	UINT m_MenuIDVolume;
